requirements.txt
streamlit==1.48.0
langchain==0.3.27
langchain_community==0.3.27
langchain_chroma==0.2.5
langchain_huggingface==0.3.1
langchain_groq==0.3.7
huggingface_hub==0.34.4
chromadb==1.0.16
pydantic==2.11.7
openpyxl==3.1.5
pandas==2.3.1
python-dotenv==1.1.1
PyPDF2==3.0.1
nest_asyncio==1.6.0
pypdf==6.0.0
sentence-transformers==5.1.0